pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vica2"
version = "0.1.0"
description = "ViCA: Visuospatial Cognitive Assistant "
requires-python = "==3.10.*" 

dependencies = [
    "asttokens==3.0.0",
    "comm==0.2.2",
    "debugpy==1.8.11",
    "decorator==5.2.1",
    "executing==2.1.0",
    "ipykernel==6.29.5",
    "ipython==8.35.0",
    "jedi==0.19.2",
    "jupyter_client==7.3.4",
    "jupyter_core==5.7.2",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pickleshare==0.7.5",
    "prompt-toolkit==3.0.51",
    "ptyprocess==0.7.0",
    "pure_eval==0.2.3",
    "pyzmq==26.2.0",
    "stack_data==0.6.3",
    "tornado==6.1",
    "traitlets==5.14.3",
    "typing_extensions==4.13.2",

    "accelerate==0.29.3",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.9.0",
    "async-timeout==5.0.1",
    "attrs==25.3.0",
    "av==14.3.0",
    "beautifulsoup4==4.13.4",
    "bitsandbytes==0.41.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "datasets==2.16.1",
    "decord==0.6.0",
    "deepspeed==0.12.2",
    "dill==0.3.7",
    "docker-pycreds==0.4.0",
    "docstring-parser==0.16",
    "einops==0.6.1",
    "einops-exts==0.0.4",
    "exceptiongroup==1.2.2",
    "fastapi==0.115.12",
    "filelock==3.18.0",
    # "flash-attn==2.5.7",
    "frozenlist==1.6.0",
    "fsspec==2023.10.0",
    "ftfy==6.3.1",
    "gdown==5.2.0",
    "gitdb==4.0.12",
    "gitpython==3.1.44",
    "gradio-client==0.2.9",
    "h11==0.14.0",
    "hf-transfer==0.1.9",
    "hjson==3.1.0",
    "httpcore==0.17.3",
    "httpx==0.24.0",
    "huggingface-hub==0.30.2",
    "hydra-core==1.3.2",
    "idna==3.10",
    "iopath==0.1.10",
    "jinja2==3.1.6",
    "joblib==1.4.2",
    "latex2mathml==3.77.0",
    # "llava==1.7.0.dev0",
    "markdown-it-py==3.0.0",
    "markdown2==2.5.3",
    "markupsafe==3.0.2",
    "mdurl==0.1.2",
    "mpmath==1.3.0",
    "multidict==6.4.3",
    "multiprocess==0.70.15",
    "networkx==3.4.2",
    "ninja==1.11.1.4",
    "numpy==1.26.1",
    "nvidia-cublas-cu12==12.1.3.1",
    "nvidia-cuda-cupti-cu12==12.1.105",
    "nvidia-cuda-nvrtc-cu12==12.1.105",
    "nvidia-cuda-runtime-cu12==12.1.105",
    "nvidia-cudnn-cu12==8.9.2.26",
    "nvidia-cufft-cu12==11.0.2.54",
    "nvidia-curand-cu12==10.3.2.106",
    "nvidia-cusolver-cu12==11.4.5.107",
    "nvidia-cusparse-cu12==12.1.0.106",
    "nvidia-ml-py==12.570.86",
    "nvidia-nccl-cu12==2.18.1",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvtx-cu12==12.1.105",
    "nvitop==1.5.0",
    "omegaconf==2.3.0",
    "open-clip-torch==2.32.0",
    "opencv-python==4.11.0.86",
    "packaging==25.0",
    "pandas==2.2.3",
    "peft==0.4.0",
    "pillow==11.2.1",
    "platformdirs==4.3.7",
    "portalocker==3.1.1",
    "propcache==0.3.1",
    "protobuf==5.29.4",
    "psutil==7.0.0",
    "py-cpuinfo==9.0.0",
    "pyarrow==19.0.1",
    "pyarrow-hotfix==0.6",
    "pydantic==1.10.8", 
    "pygments==2.19.1",
    "pynvml==12.0.0",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rich==14.0.0",
    "safetensors==0.5.3",
    "scikit-learn==1.2.2",
    "scipy==1.15.2",
    "sentencepiece==0.1.99",
    "sentry-sdk==2.26.1",
    "setproctitle==1.3.5",
    "shortuuid==1.0.13",
    "shtab==1.7.2",
    "six==1.17.0",
    "smmap==5.0.2",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "starlette==0.46.2",
    "svgwrite==1.4.3",
    "sympy==1.13.3",
    "threadpoolctl==3.6.0",
    "timm==1.0.15",
    "tokenizers==0.15.2",
    "torch==2.1.2",
    "torchaudio==2.1.2+cu121",
    "torchvision==0.16.2",
    "tqdm==4.67.1",
    "transformers@git+https://github.com/huggingface/transformers.git@1c39974a4c4036fd641bc1191cc32799f85715a4",
    "triton==2.1.0", 
    "typeguard==4.4.2",
    "tyro==0.9.19",
    "tzdata==2025.2",
    "urllib3==1.26.20",
    "uvicorn==0.34.2",
    "wandb==0.19.9",
    "wavedrom==2.0.3.post3",
    "wcwidth==0.2.13",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "yarl==1.20.0",
]


[project.urls]
# homepage = ""
repository = "https://github.com/nkkbr/ViCA"

[tool.setuptools]
packages = ["vica2", "llava", "sam2"]